api/requirements.txt
Flask==3.0.0
Pillow==10.1.0
pdf2image==1.17.0
poppler-utils